move runtime-specific code from runtime.go into separate files for better organization
